requirements.txt
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
torch==2.5.1
ucimlrepo==0.0.7
wandb==0.23.0
